+++ buildSrc/src/main/kotlin/com/danrusu/pods4k/immutableArrays/immutableArraysModule/specializations/MapSpecializationGenerator.kt
@@ -1,7 +1,6 @@
 package com.danrusu.pods4k.immutableArrays.immutableArraysModule.specializations
 
 import com.danrusu.pods4k.immutableArrays.BaseType
-import com.danrusu.pods4k.utils.emptyLine
 import com.danrusu.pods4k.utils.function
 import com.danrusu.pods4k.utils.statement
 import com.squareup.kotlinpoet.FileSpec
@@ -40,6 +39,7 @@ private fun FileSpec.Builder.addMapFunction(fromType: BaseType, toType: BaseType
             )
         },
         returns = resultTypeName,
+        forceFunctionBody = true,
     ) {
         addAnnotation(OverloadResolutionByLambdaReturnType::class)
         if (fromType == BaseType.GENERIC) {
@@ -48,7 +48,6 @@ private fun FileSpec.Builder.addMapFunction(fromType: BaseType, toType: BaseType
         if (toType == BaseType.GENERIC) {
             addTypeVariable(mappedType as TypeVariableName)
         }
-        emptyLine() // force regular function body.  Spotless formatting will remove the empty line
         statement("return ${toType.generatedClassName}(size) { transform(get(it)) }")
     }
 }